feat(deps): use choo
