+++ index.html
@@ -3,10 +3,12 @@
   <head>
     <title>Dat Desktop</title>
     <link rel=stylesheet href=./public/css/main.css>
-    <script src="https://cdn.auth0.com/js/lock/10.2/lock.js"></script>
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
   </head>
   <body>
-    <script>require('./app')</script>
+    <script
+      src="http://localhost:8080/bundle.js"
+      type="text/javascript">
+    </script>
   </body>
 </html>
