+++ index.js
@@ -1,10 +1,19 @@
 const {app, BrowserWindow, ipcMain, Menu, shell} = require('electron')
 const defaultMenu = require('electron-default-menu')
+const createServer = require('./server')
 
 const env = process.env.NODE_ENV
 
 let win, file, link
 
+// boot up the asset server
+// TODO(yw): make this run on a random port
+// TODO(yw): use a static bundle for production
+// TODO(yw): move the HTML file to a package
+// TODO(yw): listen for server callback
+createServer(8080)
+
+// handle electron events
 ipcMain.on('ready', () => {
   if (file) {
     let path = file
@@ -18,6 +27,8 @@ ipcMain.on('ready', () => {
   }
 })
 
+// create the HTML window inside the electron client
+// null -> null
 function createWindow () {
   win = new BrowserWindow({ width: 1000 })
   win.loadURL(`file://${__dirname}/index.html`)
