+++ package.json
@@ -6,8 +6,11 @@
   "repository": "juliangruber/dat-desktop",
   "description": "Dat Desktop App",
   "dependencies": {
+    "bulk-require": "^1.0.0",
     "bytes": "^2.4.0",
     "bytewise": "^1.1.0",
+    "choo": "^4.0.0-2",
+    "choo-log": "^2.0.1",
     "collect-stream": "^1.1.1",
     "dat-design": "^2.0.0",
     "dat-encoding": "^3.0.0",
@@ -16,14 +19,16 @@
     "hyperdrive": "^7.3.0",
     "hyperdrive-archive-swarm": "^3.1.0",
     "hyperdrive-import-files": "^2.2.0",
+    "insert-css": "^1.0.0",
     "js-alert": "^1.0.4",
     "level": "^1.4.0",
     "level-live-stream": "^1.4.11",
     "minimist": "^1.2.0",
     "random-access-file": "^1.2.0",
     "rimraf": "^2.5.2",
-    "yo-css": "^1.1.0",
-    "yo-yo": "^1.2.1"
+    "sheetify": "^5.1.1",
+    "tachyons": "^4.5.4",
+    "yo-css": "^1.1.0"
   },
   "scripts": {
     "start": "NODE_ENV=development electron .",
@@ -37,6 +42,9 @@
   },
   "devDependencies": {
     "appdmg": "^0.4.5",
+    "bankai": "^3.3.1",
+    "browserify": "^13.1.1",
+    "bulkify": "^1.4.2",
     "dat-icons": "^1.3.0",
     "electron-packager": "^7.0.2",
     "node-sass": "^3.8.0",
